make envs store IORefs
the code is pretty gross! the reason for introducing this is so that
variables which are closed over are given proper treatment (see
examples/counter.lox for an example. this currently has obviously
incorrect behaviour). i think this is a real turning point where you
could reasonably criticise the pervasive use of IO, but i want to follow
the book's language's semantics as closely as possible and i put a lot
of time into thinking about ways to do that without IO creeping in
everywhere, coming up with nothing. specifically, i came up with two
other candidate approaches:

a) i could use STRefs instead, which end up looking almost the same but
have the benefit of being exectued in the ST monad which is weaker than
IO. this would mean a lot of changing function signatures and doesn't
mesh very well with the current mtl-style effects architecture that i
currently have going.

b) when entering a function's evaluation, i could change the env to
include its declaration context, and then upon exiting perform a union
of the resulting environment and the environment from which it was
called. this would be quite involved and would more or less be a hack to
emulate the behaviour of mutable references, so i ended up just going
with the obvious (but gross) approach of wrapping values in IORefs.