Revert "make envs store IORefs"
This reverts commit 247c8861d082c49991360ae218fcf99371ce83db.

i spent much of yesterday trying to implement resolution like the book
describes, and ended up in debug hell with print statements everywhere,
with terrible code everywhere, with sadness everywhere. i have questions
about the soundness of the algorithm that the book presents (how are
recursive calls treated?), and i also am convinced that i don't really
like the idea of variables being closed over anyway.

why should a function call be able to modify a variable in some
disparate part of the program, in state that may or may not only exist
inside the function? if you want state local to a function, my current
thinking is that you really want encapsulation/classes. we can implement
that instead! i may revisit this later if i change my mind.

it was a reasonably helpful exercise for teaching me how to do
context-sensitive syntax checks, so i will go back and implement some
more of those.